pyproject.toml
[project]
name = "chartocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8,<3.9"
dependencies = [
    "bresenham>=0.2.1",
    "chardet>=5.2.0",
    "colorgram-py>=1.2.0",
    "matplotlib>=3.7.5",
    #"mmcv-full==1.7.2",
    #"mmdet==2.28.2",
    "mmengine>=0.10.7",
    "opencv-python<=4.6.0.66",
    "openmim>=0.3.9",
    "paddleocr>=3.3.0",
    "pillow>=10.4.0",
    "pytesseract>=0.3.13",
    "scikit-image>=0.21.0",
    "scipy==1.9.3",
    "setuptools>=75.3.2",
    "toml>=0.10.2",
    "torch==1.13.1",
    "torchvision==0.14.1",
    "tqdm>=4.67.1",
]

#[tool.uv.sources]
#mmdet = { path = "mmdetection", editable = true }



[build-system]
requires = ["setuptools>=75.3.2", "wheel"]
build-backend = "setuptools.build_meta"


[tool.setuptools.packages.find]
where = ["."]
exclude = ["mmdetection*", "data*", ".*"]
